moved declarations into single file
