--- src/types/pointer-tensor.d.ts
@@ -1,12 +1,0 @@
-import { SerdeSimplifiable } from '.';
-
-export declare class PointerTensor implements SerdeSimplifiable {
-  id: any;
-  idAtLocation: any;
-  locationId: any;
-  pointToAttr: any;
-  shape: any;
-  garbageCollectData: any;
-
-  serdeSimplify(f: any): string;
-}
